unify claude option handling
