+++ common/src/options/vertexai.ts
@@ -1,12 +1,15 @@
 import { type ModelOptionInfoItem, type ModelOptions, type ModelOptionsInfo, OptionType, SharedOptions } from "../types.js";
-import { getMaxOutputTokens } from "./context-windows.js";
 import { textOptionsFallback } from "./fallback.js";
 import {
-    getAvailableEffortLevels,
+    buildClaudeCacheOptions,
+    buildClaudeCacheTtlOptions,
+    buildClaudeEffortOptions,
+    buildClaudeThinkingOptions,
+    getClaudeMaxTokensLimit,
+} from "./shared-parsing.js";
+import {
     hasSamplingParameterRestriction,
     isGeminiModelVersionGte,
-    requiresAdaptiveThinkingOnly,
-    supportsAdaptiveThinking,
 } from "./version-parsing.js";
 
 // Union type of all VertexAI options
@@ -568,82 +571,16 @@ function getClaudeOptions(model: string, option?: ModelOptions): ModelOptionsInf
         integer: true, step: 200, description: "The maximum number of tokens to generate"
     }];
 
-    const claudeCacheOptions: ModelOptionInfoItem[] = [
-        {
-            name: "cache_enabled",
-            type: OptionType.boolean,
-            default: false,
-            description: "Enable prompt caching. Injects cache breakpoints at the system prompt, tools, and conversation pivot.",
-        },
-    ];
-    const claudeCacheTtlOptions: ModelOptionInfoItem[] = (option as VertexAIClaudeOptions)?.cache_enabled ? [
-        {
-            name: "cache_ttl",
-            type: OptionType.enum,
-            enum: { "5 minutes (default)": "5m", "1 hour": "1h" },
-            default: "5m",
-            description: "TTL for cache breakpoints. '1h' requires extended caching to be enabled on your account.",
-        }
-    ] : [];
-
-    // Check if this model supports adaptive thinking (Opus 4.6+, Sonnet 4.6+)
-    const supportsAdaptive = supportsAdaptiveThinking(model);
-    // Check if this is Opus 4.7+ where extended thinking returns 400 error
-    const adaptiveOnly = requiresAdaptiveThinkingOnly(model);
-
-    // Effort option — shown for all models that support it (Opus 4.5+, Sonnet 4.6+, all 4.7+)
-    const effortLevels = getAvailableEffortLevels(model);
-    const claudeEffortOptions: ModelOptionInfoItem[] = effortLevels ? [
-        {
-            name: "effort",
-            type: OptionType.enum,
-            enum: effortLevels,
-            default: "high",
-            description: "Controls how many tokens Claude uses when responding. Lower effort trades thoroughness for speed and cost savings.",
-        },
-    ] : [];
-
-    if (model.includes("-3-7") || supportsAdaptive) {
-        // Models with adaptive thinking support use adaptive mode with display
-        // Older models (3.7) use extended thinking (enabled/disabled)
-        const useAdaptiveThinking = supportsAdaptive;
-        // Effort is already shown via claudeEffortOptions (with xhigh/max for Opus 4.7+)
-        const claudeModeOptions: ModelOptionInfoItem[] = [];
-        const claudeThinkingOptions: ModelOptionInfoItem[] = [
-            {
-                name: "include_thoughts",
-                type: OptionType.boolean,
-                default: false,
-                description: useAdaptiveThinking
-                    ? (adaptiveOnly
-                        ? "Show the summarized thinking content in the response"
-                        : "Show the summarized thinking content in the response (default on this model)")
-                    : "Include the model's reasoning process in the response"
-            }
-        ];
-
-        return {
-            _option_id: "vertexai-claude",
-            options: [
-                ...max_tokens,
-                ...commonOptions,
-                ...claudeEffortOptions,
-                ...claudeModeOptions,
-                ...claudeThinkingOptions,
-                ...claudeCacheOptions,
-                ...claudeCacheTtlOptions,
-            ]
-        };
-    }
     return {
         _option_id: "vertexai-claude",
         options: [
             ...max_tokens,
             ...commonOptions,
-            ...claudeEffortOptions,
-            ...claudeCacheOptions,
-            ...claudeCacheTtlOptions,
-        ]
+            ...buildClaudeEffortOptions(model),
+            ...buildClaudeThinkingOptions(model),
+            ...buildClaudeCacheOptions(),
+            ...buildClaudeCacheTtlOptions((option as VertexAIClaudeOptions)?.cache_enabled),
+        ],
     };
 }
 
@@ -690,14 +627,6 @@ function getGeminiMaxTokensLimit(model: string): number {
         return 2048;
     }
     return 8192;
-}
-
-// Delegate to provider-agnostic limits,
-// override only where VertexAI supports extended output (128K for 3.7 and Opus 4.7+)
-function getClaudeMaxTokensLimit(model: string): number {
-    if (model.includes('-3-7')) return 128000;
-    if (model.includes('opus-4-7')) return 128000;
-    return getMaxOutputTokens(model);
 }
 
 function getLlamaMaxTokensLimit(_model: string): number {
